# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Auth API",
    "version": "1.0.0",
    "description": "Supabase Auth API with Protected Bearer Endpoints"
  },
  "paths": {
    "/public/info": {
      "get": {
        "summary": "Public Info",
        "responses": {
          "200": { "description": "Public data retrieved successfully" }
        }
      }
    },
    "/auth/signup": {
      "post": {
        "summary": "User Signup",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created" },
          "400": { "description": "Bad Request" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "User Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Logged in successfully" },
          "401": { "description": "Invalid credentials" }
        }
      }
    },
    "/protected/profile": {
      "get": {
        "summary": "Get Profile Data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Profile data retrieved" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/protected/dashboard": {
      "get": {
        "summary": "Get Dashboard Data",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Dashboard data retrieved" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "summary": "User Logout",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "204": { "description": "No Content" },
          "401": { "description": "Unauthorized" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}